Added proof and proof response structs
