+++ contracts/batch-signer/src/lib.rs
@@ -1,5 +1,5 @@
-pub mod batch;
 pub mod contract;
 pub mod error;
 pub mod msg;
 pub mod state;
+pub mod types;
